requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
chromadb
groq
numpy
pickle-mixin
